+++ node/src/RedisClusterClient.ts
@@ -16,6 +16,7 @@ import {
     createInfo,
     createPing,
 } from "./Commands";
+import { RequestError } from "./Errors";
 import { connection_request, redis_request } from "./ProtobufMessage";
 import { ClusterTransaction } from "./Transaction";
 
@@ -56,11 +57,11 @@ export type SlotKeyTypes = {
 export type RouteByAddress = {
     type: "routeByAddress";
     /**
-     * DNS name of the host.
+     *The endpoint of the node. If `port` is not provided, should be in the `${address}:${port}` format, where `address` is the preferred endpoint as shown in the output of the `CLUSTER SLOTS` command.
      */
     host: string;
     /**
-     * The port to access on the node. If port is not provided, `host` is assumed to be in the format `{hostname}:{port}`.
+     * The port to access on the node. If port is not provided, `host` is assumed to be in the format `${address}:${port}`.
      */
     port?: number;
 };
@@ -146,7 +147,7 @@ function toProtobufRoute(
             const split = host.split(":");
 
             if (split.length !== 2) {
-                throw new Error(
+                throw new RequestError(
                     "No port provided, expected host to be formatted as `{hostname}:{port}`. Received " +
                         host
                 );
@@ -243,7 +244,7 @@ export class RedisClusterClient extends BaseClient {
     /** Ping the Redis server.
      * See https://redis.io/commands/ping/ for details.
      *
-     * @param message - An optional message to include in the PING command. 
+     * @param message - An optional message to include in the PING command.
      * If not provided, the server will respond with "PONG".
      * If provided, the server will respond with a copy of the message.
      * @param route - The command will be routed to all primaries, unless `route` is provided, in which
@@ -251,7 +252,10 @@ export class RedisClusterClient extends BaseClient {
      * @returns - "PONG" if `message` is not provided, otherwise return a copy of `message`.
      */
     public ping(message?: string, route?: Routes): Promise<string> {
-        return this.createWritePromise(createPing(message), toProtobufRoute(route));
+        return this.createWritePromise(
+            createPing(message),
+            toProtobufRoute(route)
+        );
     }
 
     /** Get information and statistics about the Redis server.